+++ app/src/laos/java/org/eyeseetea/malariacare/strategies/PopulateDBStrategy.java
@@ -1,0 +1,7 @@
+package org.eyeseetea.malariacare.strategies;
+
+public class PopulateDBStrategy {
+
+    public static void init() {
+    }
+}
