+++ app/src/main/java/org/eyeseetea/malariacare/views/question/singlequestion/DynamicStockImageRadioButtonSingleQuestionView.java
@@ -156,8 +156,7 @@ public class DynamicStockImageRadioButtonSingleQuestionView extends LinearLayout
         //Setting a value for the stock question to get max total question correct
         View stockHideView = new View(context);
         stockHideView.setTag(Treatment.getDynamicStockQuestion());
-        Question pqHideQuestion = Question.findByUID(
-                context.getString(R.string.stockPqQuestionUID));
+        Question pqHideQuestion = QuestionStrategy.getStockPqQuestion();
         Option falseOption = Option.findById(41l);
         Value valuePq = pqHideQuestion.getValueBySession();
         if (valuePq != null) {
